Make a better code entrypoint
